Promote ebo_storage to upper namespace
ebo_storage has been removed from the 'detail' namespace and promoted
to the bit::memory namespace for better visibility
