+++ include/bit/memory/adapters/polymorphic_allocator_deleter.hpp
@@ -10,7 +10,7 @@
 #ifndef BIT_MEMORY_ADAPTERS_POLYMORPHIC_ALLOCATOR_DELETER_HPP
 #define BIT_MEMORY_ADAPTERS_POLYMORPHIC_ALLOCATOR_DELETER_HPP
 
-#include "../utilities/detail/ebo_storage.hpp"
+#include "../utilities/ebo_storage.hpp"
 #include "../utilities/pointer_utilities.hpp" // to_raw_pointer
 
 #include "../concepts/AllocatorStorage.hpp"
@@ -36,9 +36,9 @@ namespace bit {
     ///////////////////////////////////////////////////////////////////////////
     template<typename T, typename AllocatorStorage>
     class polymorphic_allocator_deleter
-      : private detail::ebo_storage<AllocatorStorage>
+      : private ebo_storage<AllocatorStorage>
     {
-      using base_type = detail::ebo_storage<AllocatorStorage>;
+      using base_type = ebo_storage<AllocatorStorage>;
       using allocator_type = typename AllocatorStorage::allocator_type;
 
       using alloc_traits   = allocator_traits<allocator_type>;
@@ -152,9 +152,9 @@ namespace bit {
 
     template<typename T, typename AllocatorStorage>
     class polymorphic_allocator_deleter<T[],AllocatorStorage>
-      : private detail::ebo_storage<AllocatorStorage>
+      : private ebo_storage<AllocatorStorage>
     {
-      using base_type = detail::ebo_storage<AllocatorStorage>;
+      using base_type = ebo_storage<AllocatorStorage>;
       using allocator_type = typename AllocatorStorage::allocator_type;
 
       using alloc_traits   = allocator_traits<allocator_type>;
