+++ include/bit/memory/allocators/fallback_allocator.hpp
@@ -11,7 +11,7 @@
 
 #include "detail/named_allocator.hpp" // detail::named_allocator
 
-#include "../utilities/detail/ebo_storage.hpp" // detail::ebo_storage
+#include "../utilities/ebo_storage.hpp" // detail::ebo_storage
 #include "../utilities/macros.hpp"             // BIT_MEMORY_UNUSED
 #include "../utilities/owner.hpp"              // owner
 
@@ -46,10 +46,10 @@ namespace bit {
     ///////////////////////////////////////////////////////////////////////////
     template<typename...AllocatorStorages>
     class fallback_allocator
-      : private detail::ebo_storage<AllocatorStorages...>
+      : private ebo_storage<AllocatorStorages...>
     {
 
-      using base_type = detail::ebo_storage<AllocatorStorages...>;
+      using base_type = ebo_storage<AllocatorStorages...>;
 
       // TODO(bitwizeshift): Support pretty-pointers by determining the common
       //                     'pointer', 'const_pointer', and 'size_type' of
