+++ include/bit/memory/block_allocators/policy_block_allocator.hpp
@@ -9,7 +9,7 @@
 #ifndef BIT_MEMORY_BLOCK_ALLOCATORS_POLICY_BLOCK_ALLOCATOR_HPP
 #define BIT_MEMORY_BLOCK_ALLOCATORS_POLICY_BLOCK_ALLOCATOR_HPP
 
-#include "../utilities/detail/ebo_storage.hpp"      // detail::ebo_storage
+#include "../utilities/ebo_storage.hpp"      // detail::ebo_storage
 #include "../utilities/allocator_info.hpp"         // allocator_info
 #include "../utilities/memory_block.hpp"           // memory_block
 #include "../utilities/owner.hpp"                  // owner
@@ -74,11 +74,11 @@ namespace bit {
              typename MemoryTracker,
              typename BasicLockable>
     class policy_block_allocator
-      : private detail::ebo_storage<BlockAllocator,MemoryTagger,MemoryTracker,BasicLockable>,
+      : private ebo_storage<BlockAllocator,MemoryTagger,MemoryTracker,BasicLockable>,
         public detail::policy_block_allocator_base<BlockAllocator>
     {
       using traits_type = block_allocator_traits<BlockAllocator>;
-      using base_type = detail::ebo_storage<BlockAllocator,
+      using base_type = ebo_storage<BlockAllocator,
                                             MemoryTagger,
                                             MemoryTracker,
                                             BasicLockable>;
